Refactor player scoring > 0 test cases
The test suite for each of the scoring possibilities is largely the
same, so we should DRY it up by moving the parts that are the same into
a single definition of the test suite.

This necessarily means the individual tests are a lot more terse, and
require reading through more of the code to understand any individual
test. However, I think it might be worth it to avoid having several
hundred more lines of code in total to say the same thing.