Fix test cases data for player scoring
I had removed the bases information, because all of the tests for a
player scoring > 0 were written to overwrite them anyways.

This was short-sighted, as the tests for a player scoring 0 specifically
do not change the bases, and thus do not overwrite them. This meant we
were trying to run the tests by passing in an invalid game state.