Test interaction of Sportsball with reducer
This uses a mock version of the reducer to allow inspecting exactly how
many times the reducer is called by Sportsball, and with what arguments.

It's important to not get too specific with these tests, as otherwise
you're testing the implementation details.

For example, if we tested whether addEntry *doesn't* call the reducer,
or tested to make sure that getEntry does, then these tests would be too
brittle. There are completely valid implementations for Sportsball that
call the reducer in either of those functions.